Moved to AggregatedFeedContentClient and add more language codes to the request
