+++ app/src/main/java/org/wikipedia/dataclient/Service.kt
@@ -501,7 +501,7 @@ interface Service {
     @GET(MW_API_PREFIX + "action=wbgetentities&props=descriptions")
     suspend fun getWikidataDescription(@Query("titles") titles: String,
                                        @Query("sites") sites: String,
-                                       @Query("languages") langCode: String): Entities
+                                       @Query("languages") langCodes: String): Entities
 
     @POST(MW_API_PREFIX + "action=wbsetclaim&errorlang=uselang")
     @FormUrlEncoded
